+++ packages/form/src/Checkbox/Checkbox.spec.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { shallow } from 'enzyme';
+import { shallow, mount, debug } from 'enzyme';
 import Checkbox from './Checkbox';
 
 describe('Checkbox', () => {
@@ -11,5 +11,15 @@ describe('Checkbox', () => {
     test('Checkbox should contain a data-qa attribute', () => {
       const component = shallow(<Checkbox qa="id-1234" />);
       expect(component.props()).toHaveProperty('data-qa', 'id-1234');
+    });
+
+    test('Checkbox should update the checked prop when is is set', () => {
+      const component = mount(<Checkbox checked={true} />);
+      expect(component.find('input[type="checkbox"]').prop('checked')).toEqual(true);
+      component.setProps({
+        checked: false,
+      });
+
+      expect(component.find('input[type="checkbox"]').prop('checked')).toEqual(false)
     });
 });