+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ## [Unreleased]
 - `TextField` Added support for colored description based on the state of the TextField
 
+### Fixed
+- `Checkbox` Use checked instead of defaultChecked
+
 ## v4.6.1 - 22-09-2020
 
 ### Fixed